+++ get_sources
@@ -48,7 +48,7 @@ rdkit_release="Release_2023_09_1"
 curl -L https://github.com/rdkit/rdkit/archive/refs/tags/$rdkit_release.tar.gz -o "$rdkit_release.tar.gz"
 
 echo "Downloading and Unpacking Privateer"
-curl -L https://github.com/glycojones/privateer/archive/refs/tags/WASM-0.3.3.tar.gz -o privateer-wasm.tar.gz
+curl -L https://github.com/glycojones/privateer/archive/refs/tags/WASM-0.3.4.tar.gz -o privateer-wasm.tar.gz
 tar xf privateer-wasm.tar.gz
 mv privateer-WASM-* privateer-wasm
